+++ lua/lazyvim/plugins/extras/lang/typescript.lua
@@ -32,10 +32,21 @@ return {
           enabled = false,
         },
         vtsls = {
+          -- explicitly add default filetypes, so that we can extend
+          -- them in related extras
+          filetypes = {
+            "javascript",
+            "javascriptreact",
+            "javascript.jsx",
+            "typescript",
+            "typescriptreact",
+            "typescript.tsx",
+          },
           settings = {
             complete_function_calls = true,
             vtsls = {
               enableMoveToFileCodeAction = true,
+              autoUseWorkspaceTsdk = true,
               experimental = {
                 completion = {
                   enableServerSideFuzzyMatch = true,
@@ -119,12 +130,6 @@ return {
           -- copy typescript settings to javascript
           opts.settings.javascript =
             vim.tbl_deep_extend("force", {}, opts.settings.typescript, opts.settings.javascript or {})
-          local plugins = vim.tbl_get(opts.settings, "vtsls", "tsserver", "globalPlugins")
-          -- allow plugins to have a key for proper merging
-          -- remove the key here
-          if plugins then
-            opts.settings.vtsls.tsserver.globalPlugins = vim.tbl_values(plugins)
-          end
         end,
       },
     },
@@ -153,8 +158,7 @@ return {
             command = "node",
             -- 💀 Make sure to update this path to point to your installation
             args = {
-              require("mason-registry").get_package("js-debug-adapter"):get_install_path()
-                .. "/js-debug/src/dapDebugServer.js",
+              LazyVim.get_pkg_path("js-debug-adapter", "/js-debug/src/dapDebugServer.js"),
               "${port}",
             },
           },