+++ lua/lazyvim/plugins/lsp/init.lua
@@ -4,8 +4,6 @@ return {
     "neovim/nvim-lspconfig",
     event = "LazyFile",
     dependencies = {
-      { "folke/neoconf.nvim", cmd = "Neoconf", config = false, dependencies = { "nvim-lspconfig" } },
-      { "folke/neodev.nvim", opts = {} },
       "mason.nvim",
       "williamboman/mason-lspconfig.nvim",
     },
@@ -75,7 +73,7 @@ return {
             -- mason = false, -- set to false if you don't want this server to be installed with mason
             -- Use this to add any additional keymaps
             -- for specific lsp servers
-            ---@type LazyKeysSpec[]
+            -- ---@type LazyKeysSpec[]
             -- keys = {},
             settings = {
               Lua = {
@@ -119,10 +117,6 @@ return {
     end,
     ---@param opts PluginLspOpts
     config = function(_, opts)
-      if LazyVim.has("neoconf.nvim") then
-        require("neoconf").setup(LazyVim.opts("neoconf.nvim"))
-      end
-
       -- setup autoformat
       LazyVim.format.register(LazyVim.lsp.formatter())
 
@@ -284,19 +278,15 @@ return {
           })
         end, 100)
       end)
-      local function ensure_installed()
+
+      mr.refresh(function()
         for _, tool in ipairs(opts.ensure_installed) do
           local p = mr.get_package(tool)
           if not p:is_installed() then
             p:install()
           end
         end
-      end
-      if mr.refresh then
-        mr.refresh(ensure_installed)
-      else
-        ensure_installed()
-      end
+      end)
     end,
   },
 }