+++ lua/lazyvim/plugins/xtras.lua
@@ -5,15 +5,25 @@ local prios = {
   ["lazyvim.plugins.extras.editor.aerial"] = 100,
   ["lazyvim.plugins.extras.editor.outline"] = 100,
   ["lazyvim.plugins.extras.editor.trouble-v3"] = 100,
+  ["lazyvim.plugins.extras.lang.typescript"] = 5,
   ["lazyvim.plugins.extras.ui.edgy"] = 2,
   ["lazyvim.plugins.extras.test.core"] = 1,
   ["lazyvim.plugins.extras.dap.core"] = 1,
 }
 
 ---@type string[]
-Config.json.data.extras = LazyVim.dedup(Config.json.data.extras)
+local extras = LazyVim.dedup(Config.json.data.extras)
 
-table.sort(Config.json.data.extras, function(a, b)
+local version = vim.version()
+local v = version.major .. "_" .. version.minor
+
+local compat = { "0_9" }
+
+if vim.tbl_contains(compat, v) then
+  extras[#extras + 1] = "lazyvim.plugins.compat.nvim-" .. v
+end
+
+table.sort(extras, function(a, b)
   local pa = prios[a] or 10
   local pb = prios[b] or 10
   if pa == pb then
@@ -25,4 +35,4 @@ end)
 ---@param extra string
 return vim.tbl_map(function(extra)
   return { import = extra }
-end, Config.json.data.extras)
+end, extras)