+++ lua/lazyvim/types.lua
@@ -1,4 +1,7 @@
+---@meta
+
 ---@class LazyVimGlobals
 vim.g = {}
 
 _G.lazyvim_docs = true
+_G.LazyVim = require("lazyvim.util")